Add files via upload
The file "index.html" is the main file for our website.
You can make changes to the website in the file "index.html".
"Planets2013.svg" is the image used on our website.
"space.mp4" is a video file that I was unable to upload onto our website. If some one else could try to upload that, that would be great.